// src/types/User.ts
import { DateContainer } from "./Timesheet";

export interface User {
    tag_group_id: string;
    po_approver: string;
    cv_attachment_id: string;
    rate: number;
    rpc_updated_by: string;
    br_approvalprocess: string;
    password: string;
    te_approver: string;
    sr_approver: string;
    departmentid: string;
    cost_currency: string;
    update_workschedule: string;
    code: string;
    timezone: string;
    cost_end_date: DateContainer;
    tb_filter_set: string;
    name: string;
    hierarchy_node_ids: string;
    po_approvalprocess: string;
    rm_filter_set: string;
    ta_filter_set: string;
    tag_start_date: DateContainer;
    hint: string;
    dr_approver: string;
    rpc_user_updated: DateContainer;
    br_approver: string;
    az_approvalprocess: string;
    pm_filter_set: string;
    tag_group_attribute_id: string;
    currency: string;
    cost_centerid: string;
    project_access_nodes: string;
    rm_approver: string;
    password_forced_change: string;
    update_tag: string;
    locked: string;
    filterset_stamp: string;
    job_codeid: number;
    payroll_code: string;
    tag_end_date: DateContainer;
    report_filter_set: string;
    km_filter_set: string;
    is_user_schedule: string;
    update_cost: string;
    az_approver: string;
    picklist_label: string;
    rm_approvalprocess: string;
    cost_lc_level: string;
    role_id: string;
    dr_approvalprocess: string;
    workschedule_workhours: string;
    te_approvalprocess: string;
    ta_approvalprocess: string;
    book_assign_stamp: string;
    id: string;
    updated: DateContainer;
    filterset_ids: string;
    cost: string;
    mfa_status: string;
    workschedule_workdays: string;
    active: number;
    externalid: string;
    ta_approver: string;
    generic: number;
    pr_approver: string;
    pb_approvalprocess: string;
    external_id: string;
    type: string;
    workscheduleid: number;
    po_filter_set: string;
    primary_filter_set: string;
    te_allowance_approvalprocess: string;
    user_locationid: string;
    account_workscheduleid: string;
    om_filter_set: string;
    ssn: string;
    acct_code: string;
    ma_filter_set: string;
    te_filter_set: string;
    te_allowance_approver: string;
    sr_approvalprocess: string;
    cost_start_date: DateContainer;
    nickname: string;
    addr: {
      addr1: string;
      addr2?: string;
      addr3?: string;
      addr4?: string;
      city: string;
      contact_id?: string;
      country: string;
      email: string;
      fax?: string;
      first: string;
      id?: string;
      last: string;
      middle?: string;
      mobile?: string;
      phone?: string;
      salutation?: string;
      state: string;
      zip: string;
    };
    bccSelf?: number;
    budgetApprovalProcess?: string;
    budgetApprover?: string;
    created?: DateContainer;
    customerId?: string;
    defaultSamlSettingsName?: string;
    line_managerid?: string;
    logintime?: DateContainer;
    mfaSetupDate?: DateContainer;
    territoryId?: string;
    week_starts?: number;
  }
  export interface WhoAmIResponse {
    User: User;
    status: string; 
  }